chore: reorganise React directory structures
