+++ frontend/next.config.js
@@ -1,4 +1,7 @@
 module.exports = {
+    images: {
+        domains: ['images.unsplash.com'], 
+    },
     async redirects() {
         return [
             {
