+++ frontend/src/app/[lang]/page.tsx
@@ -1,7 +1,7 @@
 import HeroSection from '@/components/ui/HeroSection';
 import About from '@/components/about/About';
 import Roller from '@/components/ui/Roller';
-import ItemsSection from '@/components/ui/ItemsSection';
+import ItemsSection from '@/components/items/ItemsSection';
 
 export default async function Home({ params } : {params: { lang: string };}) {
   const lang = params.lang;
